backend/experiment/ursa/pyproject.toml
[tool.poetry]
name = "ursa"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
xlrd = "^2.0.1"
bs4 = "^0.0.2"
openai = "^1.40.3"
janome = "^0.5.0"
requests = "^2.32.3"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
